Implement an rpc mechanism that lets one worker control all others.
Rewrite mon.0 key and monmap distribution in terms of this rpc.

Control files *must* pass in a random cookie to job. This is used to
prevent old job instances from making a mess of the run; that has
already happened with ceph daemons from supposedly-dead clusters
talking to unrelated clusters, so this is not just paranoia.